+++ Tests/TuistCoreTests/ValueGraph/ValueGraphTraverserTests.swift
@@ -199,6 +199,60 @@ final class ValueGraphTraverserTests: TuistUnitTestCase {
         XCTAssertEqual(got.map(\.target), [b])
     }
 
+    func test_directTargetDependencies_returnsLocalProjectTargetsOnly() {
+        // Given
+        // Project A: A1 -> A2
+        //               -> (Project B) B1
+        // Project B: B1
+        let projectA = Project.test(path: "/ProjectA", name: "ProjectA")
+        let projectB = Project.test(path: "/ProjectB", name: "ProjectB")
+        let a1 = Target.test(name: "A1")
+        let a2 = Target.test(name: "A2")
+        let b1 = Target.test(name: "B1")
+        let dependencies: [ValueGraphDependency: Set<ValueGraphDependency>] = [
+            .target(name: a1.name, path: projectA.path): Set([
+                .target(name: a2.name, path: projectA.path),
+                .target(name: b1.name, path: projectB.path),
+            ]),
+        ]
+        let targets: [AbsolutePath: [String: Target]] = [
+            projectA.path: [
+                a1.name: a1,
+                a2.name: a2,
+            ],
+            projectB.path: [
+                b1.name: b1,
+            ],
+        ]
+        // Given: Value Graph
+        let valueGraph = ValueGraph.test(path: projectA.path,
+                                         projects: [projectA.path: projectA, projectB.path: projectB],
+                                         targets: targets,
+                                         dependencies: dependencies)
+        let subject = ValueGraphTraverser(graph: valueGraph)
+
+        // Given: Graph
+        let b1Node = TargetNode.test(project: projectB, target: b1)
+        let a2Node = TargetNode.test(project: projectA, target: a2)
+        let a1Node = TargetNode.test(project: projectA, target: a1, dependencies: [a2Node, b1Node])
+
+        let graph = Graph.test(entryPath: projectA.path,
+                               entryNodes: [a1Node, a2Node, b1Node],
+                               targets: [
+                                   projectA.path: [a1Node, a2Node],
+                                   projectB.path: [b1Node],
+                               ])
+        let graphTraverser = GraphTraverser(graph: graph)
+
+        // When
+        let got = subject.directTargetDependencies(path: projectA.path, name: a1.name).sorted()
+        let gotGraph = graphTraverser.directTargetDependencies(path: projectA.path, name: a1.name).sorted()
+
+        // Then
+        XCTAssertEqual(gotGraph, got)
+        XCTAssertEqual(got.map(\.target), [a2])
+    }
+
     func test_resourceBundleDependencies_returns_an_empty_list_when_a_dependency_can_host_resources() {
         // Given
         // App -> WatchApp -> Bundle